+++ tests/test_unicode.py
@@ -4,6 +4,7 @@ from __future__ import unicode_literals
 import platform
 import tempfile
 import sys
+import time
 
 import pexpect
 import unittest
@@ -162,6 +163,15 @@ class UnicodeTests(PexpectTestCase.PexpectTestCase):
         p.sendeof()
         p.expect('▁▂▃▄▅▆▇█')
 
+    def test_readline_bin_echo(self):
+        # Test using readline() with spawnu objects. pexpect 3.2 had threw
+        # a TypeError when concatenating a bytestring to a unicode type.
+
+        # given,
+        child = pexpect.spawnu('echo', ['input', ])
+
+        # exercise,
+        assert child.readline() == 'input' + child.crlf
 
 if __name__ == '__main__':
     unittest.main()